CytokineTransputerSparse/requirements.txt
adjustText==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colourmap==1.1.21
contourpy==1.3.1
cycler==0.12.1
datazets==1.1.2
et_xmlfile==2.0.0
fcsparser==0.2.8
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
idna==3.10
ignite==1.1.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pca==2.0.9
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytorch-ranger==0.1.1
pytorch_optimizer==3.6.0
pytz==2025.2
requests==2.32.3
scatterd==1.3.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
statsmodels==0.14.4
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torch-geometric==2.6.1
torch-optimizer==0.3.0
torchaudio==2.7.0
torchsummary==1.5.1
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
